+++ dtriPMS/src/main/java/dtri/com/tw/db/pgsql/dao/WorkTypeDao.java
@@ -22,5 +22,5 @@ public interface WorkTypeDao extends JpaRepository<WorkType, Long> {
 			Pageable pageable);
 
 	// delete
-	Long deleteByWtid(Integer id);
+	Long deleteByWtid(Long id);
 }
